rename views to follow convention
